Add script to add current data to cache
If the 7-tage-inzidenz.php script isn't called for a day, the
data for that day is missing in the cache and cannot be retrieved
later. So create an option to run via systemd.timer or cron a
script daily to fetch the current data and add it to the cache.